pyproject.toml
[tool.poetry]
name = "resize-image-service"
version = "0.1.0"
description = ""
authors = ["Vitor Hideyoshi <vitor.h.n.batista@gmail.com>"]
packages = [{include = "storage_service"}]

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.0.0"
redis = "^5.0.4"
requests = "^2.32.1"
rq = "^1.16.1"
python-dotenv = "^1.0.0"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
boto3 = "^1.34.109"
python-multipart = "^0.0.9"
virustotal-python = "^1.0.2"
fastapi-utils = {extras = ["all"], version = "^0.8.0"}


[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.7.0"
coverage = "^7.5.1"
pre-commit = "^3.7.1"
poethepoet = "^0.37.0"



[tool.poe.tasks]
'run'                 = "python -m storage_service"
'run:queue'           = "python -m storage_service --queue"
'run:dev'             = "python -m storage_service --dev"
'create-hooks'        = "bash .githooks/set-hooks.sh"
'test'                = "coverage run -m unittest -v"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 79
sections=[
    "FUTURE",
    "FIRSTPARTY",
    "LOCALFOLDER",
    "THIRDPARTY",
    "STDLIB",
]

